+++ src/client-android/app/build.gradle
@@ -25,6 +25,9 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+    buildFeatures {
+        viewBinding true
+    }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
@@ -32,36 +35,23 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
-
-    dynamicFeatures = [':feature_chats_list', ':feature_chat_content']
 }
 
 dependencies {
-    api project(':jniwrapper')
+    implementation project(':jniwrapper')
+    implementation project(':core')
+    implementation project(':feature_chats_list')
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    api 'androidx.core:core-ktx:1.3.2'
-    api 'androidx.appcompat:appcompat:1.2.0'
-    api "androidx.fragment:fragment-ktx:1.3.0-rc01"
-    api 'androidx.constraintlayout:constraintlayout:2.0.4'
-    api 'com.google.android.material:material:1.2.1'
-    api 'androidx.legacy:legacy-support-v4:1.0.0'
-
-    def nav_version = "2.3.2"
-    api "androidx.navigation:navigation-fragment-ktx:$nav_version"
-    api "androidx.navigation:navigation-ui-ktx:$nav_version"
-    api "androidx.navigation:navigation-dynamic-features-fragment:$nav_version"
-
-    // Hilt
-    implementation "com.google.dagger:hilt-android:2.28-alpha"
-    kapt "com.google.dagger:hilt-android-compiler:2.28-alpha"
-    implementation 'androidx.hilt:hilt-common:1.0.0-alpha02'
-    implementation 'androidx.hilt:hilt-lifecycle-viewmodel:1.0.0-alpha02'
-    kapt 'androidx.hilt:hilt-compiler:1.0.0-alpha02'
 
     // Timber
-    api 'com.jakewharton.timber:timber:4.7.1'
+    implementation 'com.jakewharton.timber:timber:4.7.1'
 
     // Groupie
-    api "com.xwray:groupie:2.8.1"
-    api "com.xwray:groupie-kotlin-android-extensions:2.8.1"
+    implementation "com.xwray:groupie:2.9.0"
+    implementation "com.xwray:groupie-kotlin-android-extensions:2.9.0"
+
+    //Hilt
+    implementation "com.google.dagger:hilt-android:$hilt_version"
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    kapt "com.google.dagger:hilt-compiler:$hilt_version"
 }
