+++ src/client-android/app/src/main/AndroidManifest.xml
@@ -8,10 +8,10 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:name="org.melon.melonmessenger.presentation.App"
+        android:name=".App"
         android:theme="@style/Theme.MelonMessenger">
         <activity
-            android:name="org.melon.melonmessenger.presentation.MainActivity"
+            android:name=".MainActivity"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
