--- src/client-android/app/src/main/java/org/melon/melonmessenger/data/base/BaseRepository.kt
@@ -1,4 +1,0 @@
-package org.melon.melonmessenger.data.base
-
-interface BaseRepository {
-}
