--- src/client-android/app/src/main/java/org/melon/melonmessenger/domain/base/BaseUseCase.kt
@@ -1,4 +1,0 @@
-package org.melon.melonmessenger.domain.base
-
-interface BaseUseCase {
-}
