--- src/client-android/app/src/main/java/org/melon/melonmessenger/presentation/base/BaseActivity.kt
@@ -1,6 +1,0 @@
-package org.melon.melonmessenger.presentation.base
-
-import androidx.annotation.LayoutRes
-import androidx.appcompat.app.AppCompatActivity
-
-open class BaseActivity(@LayoutRes layoutRes: Int) : AppCompatActivity(layoutRes)
