--- src/client-android/app/src/main/java/org/melon/melonmessenger/presentation/base/BaseFragment.kt
@@ -1,22 +1,0 @@
-package org.melon.melonmessenger.presentation.base
-
-import android.os.Bundle
-import android.view.View
-import androidx.activity.OnBackPressedCallback
-import androidx.annotation.LayoutRes
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-
-open class BaseFragment(@LayoutRes layoutRes: Int) : Fragment(layoutRes) {
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        requireActivity().onBackPressedDispatcher.addCallback(
-            viewLifecycleOwner,
-            object : OnBackPressedCallback(true) {
-                override fun handleOnBackPressed() {
-                    // TODO: why should i use it???
-                    findNavController().popBackStack()
-                }
-            })
-    }
-}
