--- src/client-android/app/src/main/res/navigation/nav_graph.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/nav_graph"
-    app:startDestination="@id/chatsListGraph">
-
-    <include-dynamic
-        android:id="@+id/chatsListGraph"
-        app:graphResName="chats_list_nav_graph"
-        app:moduleName="feature_chats_list" />
-</navigation>
