+++ src/client-android/app/src/main/res/navigation/nav_graph_main.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/mainNavGraph"
+    app:startDestination="@id/chatsListGraph">
+
+        <include app:graph="@navigation/nav_graph_chats_list" />
+
+<!--    <fragment-->
+<!--        android:id="@+id/blankFragment"-->
+<!--        android:name="org.melon.melonmessenger.BlankFragment"-->
+<!--        android:label="fragment_blank2"-->
+<!--        tools:layout="@layout/fragment_blank2" />-->
+</navigation>
