+++ src/client-android/core/src/main/java/org/melon/core/data/base/BaseRepository.kt
@@ -1,0 +1,4 @@
+package org.melon.core.data.base
+
+interface BaseRepository {
+}
