+++ src/client-android/core/src/main/java/org/melon/core/domain/base/BaseUseCase.kt
@@ -1,0 +1,4 @@
+package org.melon.core.domain.base
+
+interface BaseUseCase {
+}
