+++ src/client-android/feature_chat_content/build.gradle
@@ -1,8 +1,8 @@
 plugins {
-    id 'com.android.dynamic-feature'
+    id 'com.android.library'
     id 'kotlin-android'
-    id 'dagger.hilt.android.plugin'
     id 'kotlin-android-extensions'
+    id 'dagger.hilt.android.plugin'
 }
 apply plugin: 'kotlin-kapt'
 
@@ -11,7 +11,6 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "org.melon.feature_chat_content"
         minSdkVersion 23
         targetSdkVersion 30
         versionCode 1
@@ -35,14 +34,13 @@ android {
 }
 
 dependencies {
-    implementation project(":app")
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'androidx.core:core-ktx:1.3.2'
-    androidTestImplementation 'androidx.annotation:annotation:1.1.0'
+    implementation project(":core")
 
-    // Hilt
-    implementation "com.google.dagger:hilt-android:2.28-alpha"
-    kapt "com.google.dagger:hilt-android-compiler:2.28-alpha"
-    implementation 'androidx.hilt:hilt-common:1.0.0-alpha02'
-    kapt 'androidx.hilt:hilt-compiler:1.0.0-alpha02'
+    //Navigation
+    implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
+    implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
+
+    //Hilt
+    implementation "com.google.dagger:hilt-android:$hilt_version"
+    kapt "com.google.dagger:hilt-compiler:$hilt_version"
 }
