+++ src/client-android/feature_chat_content/src/main/res/navigation/nav_graph_chat_content.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/chatContentGraph"
     app:startDestination="@id/chatContentFragment">
 
     <fragment
         android:id="@+id/chatContentFragment"
         android:name="org.melon.feature_chat_content.presentation.fragments.ChatContentFragment"
-        android:label="ChatContentFragment" />
+        android:label="@string/app_name" />
 </navigation>
