+++ src/client-android/feature_chats_list/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'com.android.dynamic-feature'
+    id 'com.android.library'
     id 'kotlin-android'
     id 'dagger.hilt.android.plugin'
     id 'kotlin-android-extensions'
@@ -11,7 +11,6 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "org.melon.feature_chats_list"
         minSdkVersion 23
         targetSdkVersion 30
         versionCode 1
@@ -35,15 +34,12 @@ android {
 }
 
 dependencies {
-    implementation project(":app")
+    implementation project(":core")
+    implementation project(":feature_chat_content")
+
+    //Hilt
+    implementation "com.google.dagger:hilt-android:$hilt_version"
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-
-
-    // Hilt
-    implementation "com.google.dagger:hilt-android:2.28-alpha"
-    kapt "com.google.dagger:hilt-android-compiler:2.28-alpha"
-    implementation 'androidx.hilt:hilt-common:1.0.0-alpha02'
-    kapt 'androidx.hilt:hilt-compiler:1.0.0-alpha02'
+    kapt "com.google.dagger:hilt-compiler:$hilt_version"
 }
