+++ src/client-android/feature_chats_list/src/main/java/org/melon/feature_chats_list/presentation/chats_list/ChatUi.kt
@@ -1,4 +1,4 @@
-package org.melon.feature_chats_list.presentation.models
+package org.melon.feature_chats_list.presentation.chats_list
 
 data class ChatUi(
     val chatName: String,
