+++ src/client-android/feature_chats_list/src/main/java/org/melon/feature_chats_list/presentation/chats_list/ChatsListFragment.kt
@@ -1,35 +1,33 @@
-package org.melon.feature_chats_list.presentation.fragments
+package org.melon.feature_chats_list.presentation.chats_list
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.setFragmentResultListener
-import androidx.navigation.findNavController
+import androidx.fragment.app.viewModels
+import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.xwray.groupie.GroupAdapter
-import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
-import kotlinx.android.synthetic.main.fragment_chats_list.view.*
+import com.xwray.groupie.GroupieViewHolder
+import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.android.synthetic.main.fragment_chats_list.*
+import org.melon.core.presentation.base.BaseFragment
 import org.melon.feature_chats_list.R
-import org.melon.feature_chats_list.di.DaggerChatsListComponent
-import org.melon.feature_chats_list.presentation.items.ChatItem
-import org.melon.feature_chats_list.presentation.viewmodels.ChatsListViewModel
-import org.melon.melonmessenger.presentation.base.BaseFragment
-import timber.log.Timber
-import javax.inject.Inject
+import org.melon.feature_chats_list.presentation.chat_creation.ChatCreationFragment
 
-
+@AndroidEntryPoint
 class ChatsListFragment : BaseFragment(R.layout.fragment_chats_list) {
 
-    @Inject
-    lateinit var viewModel: ChatsListViewModel
+    private val viewModel by viewModels<ChatsListViewModel>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        DaggerChatsListComponent.builder().context(requireContext()).build().inject(this)
+
         setFragmentResultListener(ChatCreationFragment.REQUEST_KEY_CHAT_CREATION) { requestKey, bundle ->
             // We use a String here, but any type that can be put in a Bundle is supported
             val result = bundle.getString(ChatCreationFragment.BUNDLE_KEY_CHAT_NAME).toString()
             viewModel.onNewChatCreated(result)
             // Do something with the result
         }
+
         super.onCreate(savedInstanceState)
     }
