--- src/client-android/feature_chats_list/src/main/res/navigation/chats_list_nav_graph.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    app:startDestination="@id/chatsListFragment">
-
-    <include-dynamic
-        android:id="@+id/chatContentGraph"
-        app:graphResName="chat_content_nav_graph"
-        app:moduleName="feature_chat_content" />
-
-    <fragment
-        android:id="@+id/chatsListFragment"
-        android:name="org.melon.feature_chats_list.presentation.fragments.ChatsListFragment"
-        android:label="ChatsListFragment" />
-    <fragment
-        android:id="@+id/chatCreationFragment"
-        android:name="org.melon.feature_chats_list.presentation.fragments.ChatCreationFragment"
-        android:label="ChatCreationFragment" />
-
-    <action
-        android:id="@+id/chatContentAction"
-        app:destination="@id/chatContentGraph" />
-    <action
-        android:id="@+id/chatCreationAction"
-        app:destination="@id/chatCreationFragment" />
-    <action
-        android:id="@+id/chatsListAction"
-        app:destination="@id/chatsListFragment"
-        app:launchSingleTop="true" />
-</navigation>
