+++ src/client-android/feature_chats_list/src/main/res/navigation/nav_graph_chats_list.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/chatsListGraph"
+    app:startDestination="@id/chatsListFragment">
+
+    <include app:graph="@navigation/nav_graph_chat_content" />
+
+    <fragment
+        android:id="@+id/chatsListFragment"
+        android:name="org.melon.feature_chats_list.presentation.chats_list.ChatsListFragment"
+        android:label="@string/app_name" />
+
+    <fragment
+        android:id="@+id/chatCreationFragment"
+        android:name="org.melon.feature_chats_list.presentation.chat_creation.ChatCreationFragment"
+        android:label="@string/app_name" />
+
+    <action
+        android:id="@+id/chatContentAction"
+        app:destination="@id/chatContentGraph" />
+    <action
+        android:id="@+id/chatCreationAction"
+        app:destination="@id/chatCreationFragment" />
+
+    <fragment
+        android:id="@+id/blankFragment"
+        android:name="org.melon.feature_chats_list.BlankFragment"
+        android:label="fragment_blank"
+        tools:layout="@layout/fragment_blank" />
+
+</navigation>
