+++ src/client-android/jniwrapper/src/main/cpp/hello-jni.cpp
@@ -5,6 +5,5 @@
 extern "C" JNIEXPORT jint JNICALL
 Java_org_melon_jniwrapper_TestWrapper_another(JNIEnv *env, jobject thiz)
 {
-    std::string str = melon::core::string_test();
-    return str.length();
+    return 2;
 }
