+++ src/client-android/settings.gradle
@@ -1,3 +1,4 @@
+include ':core'
 include ':jniwrapper'
 include ':feature_chat_content'
 include ':feature_chats_list'
